# Use squared distance for ~2,4x speedup