Current version:
```zsh
➜  colors git:(master) ✗ time ./target/release/colors 10
Saved to pic10-434125630.png
./target/release/colors 10  39,40s user 0,10s system 99% cpu 39,534 total
```

New version:
```zsh
➜  colors git:(master) time ./target/release/colors 10
Saved to pic10-1362786055.png
./target/release/colors 10  16,48s user 0,12s system 99% cpu 16,618 total
```

This change doesn't have any result on the output image, because the only use for the distances are comparisons between each other. And it doesn't matter whether you make comparisons between `a` and `b` or `a^2` and `b^2`.