@isaacg1 2017-07-21
After merging, I switched to integer math (i64) rather than floating point math (f64) and cleaned up the name of the function. Thanks for the improvement!